+++ CHANGELOG.md
@@ -4,3 +4,4 @@
 
 * Rewritten to simplify Open Graph (Facebook), Twitter Cards, and custom tags.
   Some features are incompatible with 2.x but are easy to rewrite.
+* Add shortcut helpers for easy setting of meta tags.